Fix the command_executor unit tests
Remove unused names
